Admin now able to update a course location when no students enrolled; Students enrolled can be viewed in courses list; Timesheet X button removed when button time field is disabled

+++ app/Queries/DeleteCourseByIdAndLocationId.ts
@@ -1,5 +1,5 @@
 export default function () {
   return `
-  DELETE FROM courses WHERE id = :courseId AND location_id = :locationId
+  DELETE FROM courses WHERE id = :courseId AND (:admin is TRUE OR location_id = :locationId)
 `
 }
